add user register...

+++ src/info/linsword20/todo/service/task/impl/TaskServiceImpl.java
@@ -1,14 +1,19 @@
 package info.linsword20.todo.service.task.impl;
 
 import info.linsword20.todo.bean.Task;
+import info.linsword20.todo.bean.User;
 import info.linsword20.todo.dao.task.TaskDAO;
 import info.linsword20.todo.service.task.TaskService;
 
 import java.util.List;
+import java.util.Map;
+
+import org.apache.struts2.ServletActionContext;
 
 public class TaskServiceImpl implements TaskService
 {
 	private TaskDAO taskDao;
+	private User user;
 
 	public void setTaskDao(TaskDAO taskDao)
 	{
@@ -42,25 +47,48 @@ public class TaskServiceImpl implements TaskService
 	@Override
 	public void markTask(int id)
 	{
+		user = getSessionUser();
+		Task task = findById(id);
+		if (user.getId() != task.getUser_id()) return;
 		this.taskDao.MarkTask(id);
 	}
 
 	@Override
 	public void deleteTask(Task task)
 	{
+		user = getSessionUser();
+		if (user.getId() != findById(task.getId()).getUser_id()) return;
 		this.taskDao.DeleteTask(task);
 	}
 
 	@Override
 	public Task findById(int id)
 	{
-		return this.taskDao.findById(id);
+		user = getSessionUser();
+		Task task = this.taskDao.findById(id);
+
+		if (user.getId() != task.getUser_id()) return null;
+
+		return task;
 	}
 
 	@Override
 	public void updateTask(Task task)
 	{
+		user = getSessionUser();
+
+		if (user.getId() != task.getUser_id()) return;
+
 		task.setFlag(Byte.parseByte("0"));
 		this.taskDao.updateTask(task);
 	}
+
+	private User getSessionUser()
+	{
+		Map<String, Object> session = ServletActionContext.getContext()
+				.getSession();
+		User user = (User) session.get("user");
+		return user;
+	}
+
 }
